+++ NEWS.md
@@ -5,3 +5,4 @@
 - Add a new function `read_tcga_xena`.
 * fix bugs
 - Add a new function `tcga_sample_filter`.
+- Add a new function `corr_rank`